Email Notifications
I have completed the base shell of the email notification sending. The system currently operates on a button that says "send email", upon which the function grabs the logged in user's email from the database and sends and email to that user saying their cat is outside the allowed area.